requirements.txt
flask==3.0.0
nfcpy==1.0.4

flask-cors==4.0.0
pyngrok==7.0.0

pyOpenSSL==24.0.0

